float internet payments
git-svn-id: file:///svn/bee@314 71f00a00-e368-4fde-a356-73d88c167424

+++ utils/payman/inetpay.h
@@ -2,7 +2,7 @@
 #define __INETPAY_H__
 
 
-extern int lastsum;
+extern double lastsum;
 
 int InetPayment();
